ml-service/requirements.txt
tensorflow==2.13.0
keras==2.13.1
flask==2.3.0
flask-cors==4.0.0
flask-socketio==5.3.0
numpy==1.24.3
opencv-python==4.8.0
pillow==10.0.0
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
python-socketio==5.9.0
python-dotenv==1.0.0
gunicorn==21.2.0
